pyproject.toml
[project]
name = "langgraph-a2a-server"
version = "0.1.0"
description = "A2A Server implementation for LangGraph agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "starlette>=0.27.0",
    "a2a-sdk[http-server]>=0.3.7",
]

[project.optional-dependencies]
examples = [
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.5",
]

[project.urls]
Repository = "https://github.com/5enxia/langgraph-a2a-server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/langgraph_a2a_server"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.13.2",
]
